Rename token to tokens
